+++ web/source/lib/output-path.js
@@ -1,0 +1,32 @@
+/*
+	GoToSocial
+	Copyright (C) 2021-2022 GoToSocial Authors admin@gotosocial.org
+
+	This program is free software: you can redistribute it and/or modify
+	it under the terms of the GNU Affero General Public License as published by
+	the Free Software Foundation, either version 3 of the License, or
+	(at your option) any later version.
+
+	This program is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU Affero General Public License for more details.
+
+	You should have received a copy of the GNU Affero General Public License
+	along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+"use strict";
+
+const fsSync = require("fs");
+const path = require("path");
+
+function out(name = "") {
+	return path.join(__dirname, "../assets/dist/", name);
+}
+
+if (!fsSync.existsSync(out())){
+	fsSync.mkdirSync(out(), { recursive: true });
+}
+
+module.exports = out;
